I made a big update in how is caching the Cupcake object. Now this method is capable to check if the cupcacke result is equal to one save, similar or non-match with anyone. - If is equal to one save, nothing is perfomated; - If is similar to one, only updates the parts that's diferent; - If non-match with anyone a new cupcake is created.
This change improviment in a big way in the app's performance and the number of the time that's SwiftData is request to write/delete something in the persistence store.

+++ Common/Views/Cupcake View/Services/CupcakeCacher.swift
@@ -11,10 +11,10 @@ import SwiftData
 extension CupcakeView {
     enum CupcakeCacher {
         static func caching(
-            _ cupcakes: [Cupcake.Get],
+            _ cupcakesGetted: [Cupcake.Get],
             with context: ModelContext
         ) throws {
-            for cupcake in cupcakes {
+            for cupcake in cupcakesGetted {
                 let predicate = #Predicate<Cupcake> { savedCupcake in
                     savedCupcake.id == cupcake.id
                 }
@@ -22,15 +22,24 @@ extension CupcakeView {
                 let descriptor = FetchDescriptor<Cupcake>(predicate: predicate)
                 let cupcakes = try context.fetch(descriptor)
                 
-                if !cupcakes.isEmpty {
-                    for cupcake in cupcakes {
-                        context.delete(cupcake)
+                guard (!cupcakes.isEmpty) && (cupcakes.count == 1) && (cupcakes[0].isEqual(to: cupcake)) else {
+                    if cupcakes.isEmpty {
+                        let newCupcake = Cupcake(from: cupcake)
+                        context.insert(newCupcake)
+                    } else if cupcakes.count == 1 && !cupcakes[0].isEqual(to: cupcake) {
+                        cupcakes[0].update(from: cupcake)
+                    } else if cupcakes.count > 1 {
+                        for cupcake in cupcakes {
+                            context.delete(cupcake)
+                        }
+                        let newCupcake = Cupcake(from: cupcake)
+                        context.insert(newCupcake)
                     }
+                    continue
                 }
-                
-                let newCupcake = Cupcake(from: cupcake)
-                context.insert(newCupcake)
             }
+            
+            guard context.hasChanges else { return }
             
             try context.save()
         }
